requirements.txt
crewai==0.14.0
langchain-openai==0.0.5
python-dotenv==1.0.1
typing-extensions>=4.11.0
pydantic==2.6.1
pinecone-client==3.0.2
langchain-pinecone==0.0.3
protobuf>=3.20.0,<4.0.0
sentence-transformers>=2.3.1
openai>=1.61.0
